+++ Changelog.md
@@ -34,6 +34,7 @@
 * Descending text is no longer cut off in orange welcome banner [#4377](https://github.com/diaspora/diaspora/issues/4377)
 * Adjust Facebook character limit to reality [#4380](https://github.com/diaspora/diaspora/issues/4380)
 * Restore truncated URLs when posting to Twitter [#4211](https://github.com/diaspora/diaspora/issues/4211)
+* Fix mobile search tags [#4392](https://github.com/diaspora/diaspora/issues/4392)
 
 ## Features
 * Admin: add option to find users under 13 (COPPA) [#4252](https://github.com/diaspora/diaspora/pull/4252)